# Conflicts:
#	src/main/java/de/uniks/stpmon/k/controller/encounter/EncounterOverviewController.java

+++ src/main/java/de/uniks/stpmon/k/models/MonsterAttributes.java
@@ -1,10 +1,10 @@
 package de.uniks.stpmon.k.models;
 
 public record MonsterAttributes(
-        Integer health,
-        Integer attack,
-        Integer defense,
-        Integer speed
+        Float health,
+        Float attack,
+        Float defense,
+        Float speed
 ) {
 
 }
